add exception handling to PostsController

+++ Controllers/PostsController.cs
@@ -25,8 +25,12 @@ namespace Blog.API.Controllers
 
         [HttpGet("{id}", Name="Find")]
         public ActionResult<Post> Find(int id) 
-        {
-            return Ok(_repository.Find(id));
+        {   
+            Post post = _repository.Find(id);
+
+            if(post == null) return NotFound();
+
+            return Ok(post);
         }
 
         [HttpPost]
@@ -40,6 +44,10 @@ namespace Blog.API.Controllers
         [HttpPut("{id}")]
         public ActionResult Update(int id, Post post) 
         {   
+            Post existingPost = _repository.Find(id);
+            
+            if (existingPost == null) return NotFound();
+
             _repository.Update(id, post);
 
             return NoContent();
@@ -49,6 +57,9 @@ namespace Blog.API.Controllers
         public ActionResult Patch(int id, JsonPatchDocument<Post> patchDocument) 
         {   
             Post post = _repository.Find(id);
+
+            if (post == null) return NotFound();
+
             patchDocument.ApplyTo(post, ModelState);
             _repository.Update(id, post);
       
@@ -59,6 +70,9 @@ namespace Blog.API.Controllers
         public ActionResult Delete(int id) 
         {
             Post post = _repository.Find(id);
+
+            if (post == null) return NotFound();
+
             _repository.Delete(id);
 
             return NoContent();
